fps計測値の出力の精度を上げた

+++ src/qow/framework/system/ExecuteLoop.java
@@ -17,7 +17,7 @@ public class ExecuteLoop extends Loop {
      *
      * @param rate レートの初期値
      */
-    protected ExecuteLoop(long rate) {
+    protected ExecuteLoop(int rate) {
         super(rate);
     }
